requirements.txt
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.51
langchain_deepseek==0.1.3
langchain_huggingface==0.1.2
langchain_openai==0.3.12
langchain_text_splitters==0.3.8
langgraph==0.3.29
llama_index==0.12.30
python-dotenv==1.1.0
